fix: align export with graphql convention
BREAKING CHANGE: All imports need to be `import {graphql}` instead of
`import graphql` from now on

Closes #4
